Added function to the GitManager class to identify if the current commit is the initial commit so we won't perform a git reset --hard in that case

+++ scripts/pre-commit.js
@@ -62,8 +62,12 @@ if (!gitRoot) {
         exitCode = 1;
       }
 
-      command = "git reset --hard";
-      execCommand(command, gitRoot);
+      // Only perform the git reset hard if we have an initial commit on this branch.
+      // Fix: https://github.com/kazazor/git-pre-commit/issues/8
+      if (gitManager.isInitialCommitExists()) {
+        command = "git reset --hard";
+        execCommand(command, gitRoot);
+      }
 
       command = "git stash pop --quiet --index";
       execCommand(command, gitRoot);
